Add remove item, duplicate item functionality; fix migrations

+++ migrations/20191022201822_year_upgrade.js
@@ -2,7 +2,7 @@ exports.up = function( knex, Promise ) {
 	return Promise.all( [
 		knex.schema.hasTable( 'years' ).then( ( exists ) => {
 			if ( ! exists ) {
-				knex.schema.createTable( 'years', ( table ) => {
+				return knex.schema.createTable( 'years', ( table ) => {
 					table.string( 'year' ).primary();
 					table.text( 'info' );
 				} );
@@ -10,7 +10,7 @@ exports.up = function( knex, Promise ) {
 		} ),
 		knex.schema.hasTable( 'years_to_users' ).then( ( exists ) => {
 			if ( ! exists ) {
-				knex.schema.createTable( 'years_to_users', ( table ) => {
+				return knex.schema.createTable( 'years_to_users', ( table ) => {
 					table.increments( 'id' );
 					table.integer( 'user_id' ).unsigned().notNullable().references( 'users.id' ).onDelete( 'CASCADE' );
 					table.string( 'year' ).notNullable().references( 'years.year' ).onDelete( 'CASCADE' );
@@ -18,9 +18,11 @@ exports.up = function( knex, Promise ) {
 			}
 		} ),
 		knex.schema.alterTable( 'users', ( table ) => {
-			knex.schema.hasColumn( 'users', 'year_viewing').then( ( exists ) => {
-				table.dropColumn( 'year_viewing' );
-			} );	
+			return knex.schema.hasColumn( 'users', 'year_viewing').then( ( exists ) => {
+				if ( exists ) {
+					return table.dropColumn( 'year_viewing' );
+				}
+			} );
 		} ),
 		knex.schema.alterTable( 'items', ( table ) => {
 			if ( process.env.NODE_ENV && process.env.NODE_ENV === 'production' ) {
